refactor UserManager
