--- application/src/main/java/team/themoment/officialgsm/common/util/UserUtil.java
@@ -1,8 +1,0 @@
-package team.themoment.officialgsm.common.util;
-
-import team.themoment.officialgsm.domain.user.User;
-
-public interface UserUtil {
-
-    User getCurrentUser();
-}
