+++ application/src/main/java/team/themoment/officialgsm/domain/auth/usecase/ApprovedUseCase.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 import team.themoment.officialgsm.common.annotation.UseCaseWithTransaction;
 import team.themoment.officialgsm.common.exception.CustomException;
 import team.themoment.officialgsm.common.exception.CustomHttpStatus;
-import team.themoment.officialgsm.common.util.UserUtil;
 import team.themoment.officialgsm.domain.user.User;
 import team.themoment.officialgsm.repository.user.UserRepository;
 
@@ -16,10 +15,8 @@ import java.time.LocalDateTime;
 @RequiredArgsConstructor
 public class ApprovedUseCase {
     private final UserRepository userRepository;
-    private final UserUtil userUtil;
 
-    public void execute(String oauthId) {
-        User grantor = userUtil.getCurrentUser();
+    public void execute(String oauthId, User grantor) {
         User user = userRepository.findByOauthId(oauthId)
                 .orElseThrow(() -> new CustomException("유저를 찾을 수 없습니다.", CustomHttpStatus.NOT_FOUND));
         LocalDateTime approvedAt = LocalDateTime.now();
