+++ application/src/main/java/team/themoment/officialgsm/domain/auth/usecase/FindUserInfoUseCase.java
@@ -3,7 +3,6 @@ package team.themoment.officialgsm.domain.auth.usecase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import team.themoment.officialgsm.common.annotation.UseCaseWithReadOnlyTransaction;
-import team.themoment.officialgsm.common.util.UserUtil;
 import team.themoment.officialgsm.domain.auth.dto.UserInfoDto;
 import team.themoment.officialgsm.domain.user.User;
 
@@ -11,10 +10,8 @@ import team.themoment.officialgsm.domain.user.User;
 @UseCaseWithReadOnlyTransaction
 @RequiredArgsConstructor
 public class FindUserInfoUseCase {
-    private final UserUtil userUtil;
 
-    public UserInfoDto execute() {
-        User user = userUtil.getCurrentUser();
+    public UserInfoDto execute(User user) {
         return UserInfoDto.builder()
                 .userName(user.userName())
                 .userEmail(user.userEmail())
