+++ presentation/build.gradle
@@ -11,6 +11,9 @@ dependencies {
     /* spring app */
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
+    /* security */
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
     /* validation */
     implementation 'org.springframework.boot:spring-boot-starter-validation'
