+++ presentation/src/main/java/team/themoment/officialgsm/admin/auth/manager/UserManager.java
@@ -1,21 +1,18 @@
-package team.themoment.officialgsm.global.security.util;
+package team.themoment.officialgsm.admin.auth.manager;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import team.themoment.officialgsm.common.exception.CustomException;
 import team.themoment.officialgsm.common.exception.CustomHttpStatus;
-import team.themoment.officialgsm.common.util.UserUtil;
 import team.themoment.officialgsm.domain.user.User;
 import team.themoment.officialgsm.repository.user.UserRepository;
 
 @Component
 @RequiredArgsConstructor
-public class UserUtilImpl implements UserUtil {
+public class UserManager {
     private final UserRepository userRepository;
 
-    @Override
     public User getCurrentUser() {
         String oauthId = SecurityContextHolder.getContext().getAuthentication().getName();
         return userRepository.findByOauthId(oauthId)
